split eval and train into 2 files
